Add files via upload
The objective of this program is to implement vectorized approach for a softmax classifier.
The pre-processed fashion_mnist data is loaded. The number of training and testing instances were found to be 60000 and 10000 respectively with 784 features and 10 different classes. Since the objective is to build softmax classifier, the number of neurons in the layer will be equal to the number of target classes present in the dataset (10 classes in this case).  The weights are randomly initialized in the shape of (10,784) and bias is initialized to zeroes in shape (1,10). The data is fed into a forward pass function which calculates the probabilities of the output. This output is then fed into cross-entropy along with true labels which calculates the cross-entropy loss for the data. Using this loss, we calculate the derivatives and update the weights and bias. The automatic differentiator gradient tape from tensorflow is used to update the parameters using Adam optimizer. The output of forward pass is also fed into calculate accuracy function with true labels to determine the accuracy of the model. The model is trained on 100 epochs. The resulting accuracy for train data and test data is 0.77615 and 0.7663 respectively.